toycathon-master/requirements.txt
aiodns==3.0.0
aiosmtpd==1.4.2
asgiref==3.5.0
atpublic==3.0.1
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
Django==4.0.2
Django-Verify-Email==2.0.3
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
psycopg2==2.9.3
pycares==4.1.2
pycparser==2.21
pymongo==4.0.1
pytz==2021.3
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.2.4
tzdata==2021.5
verify-email==2.4.3
